requirements.txt
# Flask Web Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.1.3

# Datenbank & ORM
SQLAlchemy==2.0.40
sqlalchemy-utils==0.41.0
psycopg2-binary==2.9.9

# AI & externe APIs
requests==2.32.3
python-dotenv==1.1.0
omdb==0.3.0
google-generativeai>=0.8.0

# Testing
pytest==8.3.5

# Sicherheit & Performance
bleach==6.1.0
gunicorn==21.2.0

# Caching (optional)
redis==5.0.0

# Logging
colorlog==6.8.0

# Passwort-Hashing
bcrypt==4.1.2

# Zeitzone-Unterstützung
pytz==2024.1

# Entwicklungstools (optional)
flask-cors==4.0.0
